pankaj sharma
A leap year is a year in which an extra day is added to the Gregorian calendar, which is used by most of the world. While an ordinary year has 365 days, a leap year has 366 days. ... In an ordinary year, February has 28 days, but in a leap year, it has 29 days.