Create Stairs Pattern
A staircase is a simple java pattern program which requires the knowledge of how loops work in Java. In this program, we use the concept of nested for-loops to print the pattern. The outer for loop is used to iterate over the height of the staircase and the inner for loop is used to print the pattern.